Remove Blobject from Swift
