+++ swift/test/Ice/objects/Server.swift
@@ -21,7 +21,7 @@ class Server: TestHelperI {
         try adapter.add(servant: InitialDisp(InitialI(adapter)), id: Ice.stringToIdentity("initial"))
         try adapter.add(servant: F2Disp(F2I()), id: Ice.stringToIdentity("F21"))
         try adapter.add(
-            servant: BlobjectDisp(UnexpectedObjectExceptionTestI()),
+            servant: UnexpectedObjectExceptionTestI(),
             id: Ice.stringToIdentity("uoet"))
         try adapter.activate()
         serverReady()
